requirements.txt
flask
flask-bcrypt
flask-jwt-extended
flask-cors
mysql-connector-python
gunicorn